Removing PluginBase classes in favour of simple interface thingies
