+++ src/Noobot.Core/DependencyResolution/ContainerFactory.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using Noobot.Core.Configuration;
 using Noobot.Core.MessagingPipeline;
 using Noobot.Core.Plugins;
+using Noobot.Core.Plugins.StandardPlugins;
 using Noobot.Core.Slack;
 using StructureMap.Configuration.DSL;
 using StructureMap.Graph;
@@ -11,19 +13,19 @@ namespace Noobot.Core.DependencyResolution
     public class ContainerFactory : IContainerFactory
     {
         private readonly IPipelineManager _pipelineManager;
-        private readonly IPluginManager _pluginManager;
+        private readonly IPluginConfiguration _pluginConfiguration;
 
-        private readonly Type[] _singletons = 
+        private readonly Type[] _singletons =
         {
             typeof(INoobotCore),
             typeof(IPipelineFactory),
             typeof(IConfigReader),
         };
 
-        public ContainerFactory(IPipelineManager pipelineManager, IPluginManager pluginManager)
+        public ContainerFactory(IPipelineManager pipelineManager, IPluginConfiguration pluginConfiguration)
         {
             _pipelineManager = pipelineManager;
-            _pluginManager = pluginManager;
+            _pluginConfiguration = pluginConfiguration;
         }
 
         public INoobotContainer CreateContainer()
@@ -36,21 +38,50 @@ namespace Noobot.Core.DependencyResolution
                 x.WithDefaultConventions();
             });
 
-            registry = _pipelineManager.Initialise(registry);
-            registry = _pluginManager.Initialise(registry);
-
             foreach (Type type in _singletons)
             {
                 registry.For(type).Singleton();
             }
 
-            Type[] pluginTypes = _pluginManager.ListPluginTypes();
+            registry = _pipelineManager.Initialise(registry);
+            Type[] pluginTypes = SetupPlugins(registry);
+
             var container = new NoobotContainer(registry, pluginTypes);
 
             IPipelineFactory pipelineFactory = container.GetInstance<IPipelineFactory>();
             pipelineFactory.SetContainer(container);
 
             return container;
+        }
+
+        private Type[] SetupPlugins(Registry registry)
+        {
+            var pluginTypes = new List<Type>(_pluginConfiguration.ListPluginTypes())
+            {
+                typeof (StoragePlugin),
+                typeof (SchedulePlugin),
+                typeof (StatsPlugin),
+                typeof (AdminPlugin)
+            };
+
+            registry.Scan(x =>
+            {
+                // scan assemblies that we are loading pipelines from
+                foreach (Type pluginType in pluginTypes)
+                {
+                    x.AssemblyContainingType(pluginType);
+                }
+            });
+
+            // make all plugins singletons
+            foreach (Type pluginType in pluginTypes)
+            {
+                registry
+                    .For(pluginType)
+                    .Singleton();
+            }
+
+            return pluginTypes.ToArray();
         }
     }
 }
