+++ src/Noobot.Core/Plugins/IPluginConfiguration.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Noobot.Core.Plugins
+{
+    public interface IPluginConfiguration
+    {
+        Type[] ListPluginTypes();
+    }
+}
