--- src/Noobot.Core/Plugins/IPluginManager.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using StructureMap.Configuration.DSL;
-
-namespace Noobot.Core.Plugins
-{
-    public interface IPluginManager
-    {
-        Registry Initialise(Registry registry);
-        Type[] ListPluginTypes();
-    }
-}
