--- src/Noobot.Custom/PluginManager.cs
@@ -1,13 +1,0 @@
-﻿using Noobot.Core.Plugins;
-using Noobot.Custom.Plugins;
-
-namespace Noobot.Custom
-{
-    public class PluginManager : PluginManagerBase
-    {
-        protected override void Initialise()
-        {
-            Use<PingPlugin>();
-        }
-    }
-}
